pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "shadow9-manager"
version = "1.0.0"
description = "Secure SOCKS5 proxy with Tor network support"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Shadow9 Team"}
]
keywords = ["socks5", "proxy", "tor", "privacy", "security"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Internet :: Proxy Servers",
    "Topic :: Security",
]

dependencies = [
    # Core async networking
    "aiohttp>=3.9.0",
    "aiohttp-socks>=0.8.4",
    "PySocks>=1.7.1",
    # Security and encryption
    "cryptography>=41.0.0",
    "argon2-cffi>=23.1.0",
    # Configuration
    "pyyaml>=6.0.1",
    "python-dotenv>=1.0.0",
    # CLI
    "typer>=0.9.0",
    "rich>=13.7.0",
    # Logging
    "structlog>=24.1.0",
    # REST API (FastAPI)
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    # Strong typing (Pydantic v2)
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "black>=24.0.0",
    "mypy>=1.8.0",
]

[project.scripts]
shadow9 = "shadow9.cli:cli"
shadow9-api = "shadow9.api.app:run_server"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ['py310', 'py311', 'py312']

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
